GAP and padding tweak
Added a new value for GAP, this required an adjustment of the other GAP values to accommodate for the new GAP.lg value. The number values should be the same, the reference is just updated. The vertical padding has also been removed. The new GAP.lg value makes this padding not needed.