GAP and cornerRadius and padding tweaks
Added a new value for GAP, this required an adjustment of the other GAP values to accommodate for the new GAP.lg value. The number values should be the same, the reference is just updated. The vertical padding has also been removed for last row. The new GAP.lg value makes this last row padding not needed. The radius tweak rounds the corners when avatars are hidden and squares off the top when shown to make it flush with the circle.