Divider Corner Radius
very small adjustment to make dividers have rounded corners. All dividers, lines, and dashed lines should be rounded.

The only place dashes are not rounded is the log type button. This will require a bit of refactor that isn't as easy to just get in line like the rest.

I switched out the Rectangle for a Line because of how the dashes work.