fixes empty changelog and link count bug
if user enters some text for a changelog and then wants to delete it, they can't. This is fixed.

users were allowed to add more than 8 links. This revealed a larger issue in terms of where links were being moved and referenced. Should be fixed now. Testing with the beta widget now.